Be more forgiving with available CPU scaling governors assert
Rather than enforcing that all of the device's scaling governors are
in the KNOWN_CPU_SCALING_GOVERNOR_VALUES list, enforce that at least
one of them is in the list. We already enforce that the desired governor
is in the known list when we try to set it, so we just need to ensure
that there's at least one available which is valid.