python-adb: Make FindDevicesSafe true to its name.
Previously, if FindDevices threw an exception, FindDevicesSafe would
catch it and proceed to poll FindDevices for more devices. But if a
python generator throws an exception, it stops all yielding. So it
wasn't really finding all devices.

This just copy pastes the code in FindDevices into FindDevicesSafe
with some try-catching'ing.

BUG=670879
R=dpranke@chromium.org

Review-Url: https://codereview.chromium.org/2615913002 .